requirements.txt
# Core dependencies
numpy>=2.1.0
pandas>=2.2.0
scipy>=1.15.0
scikit-learn>=1.6.0

# PyTorch and related (Python 3.13 compatible versions)
torch>=2.6.0
torchvision>=0.21.0
pytorch-lightning>=2.4.0
torchmetrics>=1.5.0

# ML/DL frameworks
transformers>=4.46.0
datasets>=3.1.0
accelerate>=1.1.0

# Visualization and data processing
matplotlib==3.9.2
seaborn==0.13.2
polars==1.14.0
plotly

# Configuration and utilities
hydra-core==1.3.2
omegaconf==2.3.0
pyyaml==6.0.2
pydantic==2.9.2

# Data and file handling
pyarrow==18.0.0
pyfaidx==0.8.1.3

# Specialized ML utilities
lightly>=1.5.0
einops>=0.8.0
enformer-pytorch>=0.8.0
vector-quantize-pytorch>=1.20.0

# Development and monitoring
wandb==0.18.7
ipython==8.36.0

# Pip dependencies from environment.yml
aenum==3.1.15
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
click==8.1.7
decorator==5.2.1
dill==0.3.8
discrete-key-value-bottleneck-pytorch==0.1.1
docker-pycreds==0.4.0
einx==0.3.0
exceptiongroup==1.3.0
executing==2.2.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
gitpython==3.1.43
huggingface-hub==0.26.2
importlib-metadata==8.5.0
jedi==0.19.2
lightly-utils==0.0.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt-toolkit==3.0.51
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
regex==2024.11.6
rich==13.9.4
safetensors>=0.5.0
sentry-sdk>=2.18.0
setproctitle>=1.3.4
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.20.3
torchinfo==1.8.0
traitlets==5.14.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.17.2
zipp==3.21.0
